Import OpenAFS 1.3.73

+++ src/WINNT/afsd/cm_buf.h
@@ -72,7 +72,7 @@ typedef struct cm_buf {
 				 */
         struct cm_buf *allp;	/* next in all list */
 	osi_mutex_t mx;		/* mutex protecting structure except refcount */
-    int refCount;		/* reference count (buf_globalLock) */
+    unsigned long refCount;		/* reference count (buf_globalLock) */
         long idCounter;		/* counter for softrefs; bumped at each recycle */
         long dirtyCounter;	/* bumped at each dirty->clean transition */
 #ifdef notdef
